CPSC532-HW1/wandb/run-20220920_140727-1gmtewa0/files/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
antlr4-python3-runtime==4.9.3
anyio==3.3.1
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.1.0
arviz==0.11.4
astor==0.8.1
async-timeout==4.0.2
attrs==21.2.0
autograd==1.4
babel==2.9.1
backcall==0.2.0
bambi==0.7.1
bleach==4.1.0
cachetools==5.0.0
certifi==2021.5.30
cffi==1.14.6
cftime==1.5.2
charset-normalizer==2.0.6
click-plugins==1.1.1
click==8.0.1
cligj==0.7.2
cloudpickle==2.0.0
covidcast==0.1.5
cvxpy==1.1.18
cvxpylayers==0.1.5
cycler==0.10.0
databricks-cli==0.16.2
datetime==4.3
debugpy==1.4.3
decorator==5.1.0
deepdow==0.2.2
defusedxml==0.7.1
delphi-epidata==0.3.20
deprecation==2.1.0
descartes==1.1.0
diffcp==1.0.19
dill==0.3.4
docker-pycreds==0.4.0
docker==5.0.3
ecos==2.0.10
eikon==1.1.12
emcee==3.1.1
entrypoints==0.3
epiweeks==2.1.4
et-xmlfile==1.1.0
etils==0.7.1
fastprogress==1.0.2
filelock==3.6.0
fiona==1.8.21
flask-login==0.5.0
flask-sqlalchemy==2.5.1
flask==2.0.2
formulae==0.2.0
frozenlist==1.3.0
future==0.18.2
geopandas==0.10.2
gitdb==4.0.9
gitpython==3.1.26
google-auth-oauthlib==0.4.6
google-auth==2.5.0
greenlet==1.1.2
grpcio==1.43.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.7.0
httpcore==0.13.7
httpx==0.19.0
hydra-core==1.2.0
idna==3.2
imageio-ffmpeg==0.4.7
imageio==2.19.3
importlib-metadata==4.10.1
importlib-resources==5.9.0
ipykernel==6.4.1
ipython-genutils==0.2.0
ipython==7.27.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-console==6.4.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.8.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-server==1.11.0
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.2
jupyterlab==3.1.13
kiwisolver==1.3.2
llvmlite==0.38.1
lxml==4.6.3
mako==1.1.6
markdown==3.3.6
markupsafe==2.0.1
matplotlib-inline==0.1.3
matplotlib==3.4.3
mistune==0.8.4
mlflow==1.23.0
multidict==6.0.2
munch==2.5.0
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netcdf4==1.5.8
notebook==6.4.4
numba==0.55.2
numexpr==2.8.1
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.2.3
openpyxl==3.0.9
opt-einsum==3.3.0
osqp==0.6.2.post5
packaging==21.3
pandas==1.3.3
pandasgui==0.2.13
pandocfilters==1.5.0
parso==0.8.2
particles==0.3
pathtools==0.1.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.3.2
pip==22.2.2
plotly==5.3.1
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.7
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.19.3
psutil==5.9.2
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pybind11==2.9.0
pycparser==2.20
pyemd==0.5.1
pygments==2.10.0
pymc3==3.11.4
pymssql==2.2.5
pynput==1.7.3
pyobjc-core==7.3
pyobjc-framework-cocoa==7.3
pyobjc-framework-quartz==7.3
pyodbc==4.0.32
pyparsing==2.4.7
pyproj==3.3.1
pypyodbc==1.3.6
pyqt5-qt5==5.15.2
pyqt5-sip==12.9.0
pyqt5==5.15.4
pyqtwebengine-qt5==5.15.2
pyqtwebengine==5.15.4
pyreadr==0.4.4
pyro-api==0.1.2
pyro-ppl==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.3.0
qdldl==0.1.5.post0
qgrid==1.3.1
qtconsole==5.3.0
qtpy==2.1.0
qtstylish==0.1.5
querystring-parser==1.2.4
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
requests==2.26.0
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0
scipy==1.7.1
scs==3.1.0
seaborn==0.11.2
semver==2.13.0
send2trash==1.8.0
sentry-sdk==1.9.8
setproctitle==1.3.2
setuptools==57.4.0
shapely==1.8.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sqlalchemy==1.4.25
sqlparse==0.4.2
statsmodels==0.13.2
tables==3.7.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
terminado==0.12.1
testpath==0.5.0
theano-pymc==1.1.2
threadpoolctl==2.2.0
torch==1.12.1
torchdiffeq==0.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
werkzeug==2.0.2
wheel==0.37.0
widgetsnbextension==3.5.1
wordcloud==1.8.1
xarray==0.21.1
xlsxwriter==3.0.1
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0